Runtime update permission session

+++ resources/views/payroll/bonus_type.blade.php
@@ -9,6 +9,9 @@
             <span class="pull-right">
               <a v-on:click="modal_open('#bonus_type_modal'),bonusType = []" onclick="document.getElementById('bonus_type_modal_form').reset()" class="btn btn-sm btn-dark btn-gradient dark"><span class="glyphicons glyphicon-pencil"></span> &nbsp; Add Bonus Type</a>
             </span>
+            <?php 
+              $chkUrl = \Request::segment(1);
+            ?>
         </div>
     </div>
     <div class="panel-body pn">
@@ -54,15 +57,19 @@
                    <td v-text="bonusType.created_at"></td>
                    <td v-text="bonusType.updated_at"></td>
                    <td>
-                      <div class="btn-group">
+                      @if(in_array($chkUrl."/edit", session('userMenuShare')))
+                        <div class="btn-group">
                            <a v-on:click="editBonusType(bonusType.id, index, '#bonus_type_modal'),bonusType=[]" class="btn btn-sm btn-primary"><i class="glyphicons glyphicons-pencil"></i>
                            </a>
-                       </div>
-                       <div class="btn-group">
+                        </div>
+                      @endif
+                      @if(in_array($chkUrl."/delete", session('userMenuShare')))
+                        <div class="btn-group">
                            <a v-on:click="deleteBonusType(bonusType.id,index)" class="btn btn-sm btn-danger">
                                <i class="glyphicons glyphicons-bin"></i>
                            </a>
-                       </div>
+                        </div>
+                      @endif
                    </td>
                 </tr>
             </tbody>
